refactor: domain, dto layer 분리
