+++ src/main/java/roomescape/dao/ReservationUpdatingDao.java
@@ -1,14 +1,12 @@
 package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 import roomescape.domain.Reservation;
-import roomescape.domain.dto.ReservationAddRequest;
 
 @Repository
 public class ReservationUpdatingDao {
