+++ src/main/java/roomescape/service/ReservationService.java
@@ -10,8 +10,8 @@ import roomescape.dao.TimeQueryingDao;
 import roomescape.dao.TimeUpdatingDao;
 import roomescape.domain.Reservation;
 import roomescape.domain.Time;
-import roomescape.domain.dto.ReservationAddRequest;
-import roomescape.domain.dto.TimeAddRequest;
+import roomescape.dto.ReservationAddRequest;
+import roomescape.dto.TimeAddRequest;
 import roomescape.exception.InvalidReservationException;
 import roomescape.exception.InvalidTimeException;
 import roomescape.exception.NotFoundReservationException;
